admin_account
linked the account page to admin and users. so when users log-in then they go to user account and if admin log-in then they saw admin account.

+++ database/migrations/2026_02_19_211029_add_user_columns_to_vouchers_table.php
@@ -8,14 +8,41 @@ return new class extends Migration{
     public function up(): void
     {
         Schema::table('vouchers', function (Blueprint $table){
-            $table->foreignId('user_id')->nullable()->after('id')->constrained()->nullOnDelete();
-            $table->timestamp('used_at')->nullable()->after('active');
+            if(!Schema::hasColumn('vouchers', 'user_id')){
+            $table->foreignId('user_id')
+            ->nullable()
+            ->after('id')
+            ->constrained()
+            ->nullOnDelete();
+            }
+
+
+            if(!Schema::hasColumn('vouchers', 'used_at')){
+            $table->timestamp('used_at')
+            ->nullable()
+            ->after('active');
+        }
         });
+        
     }
+
+
     public function down(): void{
         Schema::table('vouchers', function (Blueprint $table){
-            $table->dropConstrainedForeignId('user_id');
+        if(Schema::hasColumn('vouchers','user_id')){
+            try
+            {
+                $table->dropConstrainedForeignId('user_id');
+            }
+            catch(\Throwable $e){
+                $table->dropColumn('user_id');
+            }
+        }
+        
+        if(Schema::hasColumn('vouchers', 'used_at')){
             $table->dropColumn('used_at');
+        }
+           
         });
     }
 };
